move clipboard copy input to it's own file
