+++ app/javascript/packs/application.js
@@ -36,3 +36,4 @@ import '../src/js/audio/navbar_tabs.js';
 import "../src/js/audio/timecode_in_anchor.js";
 import "../src/js/audio/share_link.js";
 import '../src/js/audio/jump_to_text.js';
+import "../src/js/audio/clipboard_copy_input.js";
